+++ commands/stop.yaml
