+++ global.go
@@ -1,0 +1,3 @@
+package goradbot
+
+var prefix = "!rad"
