+++ go.mod
@@ -1,0 +1,13 @@
+module goradbot
+
+go 1.14
+
+require (
+	github.com/bwmarrin/discordgo v0.20.3
+	github.com/ghodss/yaml v1.0.0 // indirect
+	github.com/gin-gonic/gin v1.6.3
+	github.com/heroku/x v0.0.24
+	github.com/jonas747/dca v0.0.0-20200403191804-067b814772f1
+	github.com/jonas747/ogg v0.0.0-20161220051205-b4f6f4cf3757 // indirect
+	gopkg.in/yaml.v2 v2.3.0 // indirect
+)
