+++ go.mod.bak
@@ -1,0 +1,3 @@
+module github.com/PAWA-cl/goradbot
+
+go 1.14
