--- goradbot/global.go
@@ -1,3 +1,0 @@
-package goradbot
-
-var prefix = "!rad"
