--- goradbot/listeners.go
@@ -1,37 +1,0 @@
-package goradbot
-
-import (
-	"fmt"
-
-	"github.com/bwmarrin/discordgo"
-)
-
-// This function will be called (due to AddHandler above) when the bot receives
-// the "ready" event from Discord.
-func ready(s *discordgo.Session, event *discordgo.Ready) {
-	// Set the playing status.
-	s.UpdateStatus(0, prefix)
-}
-
-// This function will be called (due to AddHandler above) every time a new
-// message is created on any channel that the autenticated bot has access to.
-func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-	checkValidCommand(s, m)
-}
-
-// This function will be called (due to AddHandler above) every time a new
-// guild is joined.
-func guildCreate(s *discordgo.Session, event *discordgo.GuildCreate) {
-
-	if event.Guild.Unavailable {
-		return
-	}
-
-	for _, channel := range event.Guild.Channels {
-		if channel.ID == event.Guild.ID {
-			_, _ = s.ChannelMessageSend(channel.ID,
-				fmt.Sprintf("GoRadBot is ready! Type %s while in a voice channel to play a sound.", prefix))
-			return
-		}
-	}
-}
