--- goradbot/start.go
@@ -1,57 +1,0 @@
-package goradbot
-
-import(
-	"fmt"
-	"os"
-	"os/signal"
-	"syscall"
-
-	"github.com/bwmarrin/discordgo"
-)
-
-//Start initialize the bot.
-func Start(token string){
-	if token == "" {
-		fmt.Println("No token provided. Please run: airhorn -t <bot token>")
-		return
-	}
-
-	// Load the sound file.
-	//err := loadSound()
-	//if err != nil {
-	//	fmt.Println("Error loading sound: ", err)
-	//	fmt.Println("Please copy $GOPATH/src/github.com/bwmarrin/examples/airhorn/airhorn.dca to this directory.")
-	//	return
-	//}
-
-	// Create a new Discord session using the provided bot token.
-	dg, err := discordgo.New("Bot " + token)
-	if err != nil {
-		fmt.Println("Error creating Discord session: ", err)
-		return
-	}
-
-	// Register ready as a callback for the ready events.
-	dg.AddHandler(ready)
-
-	// Register messageCreate as a callback for the messageCreate events.
-	dg.AddHandler(messageCreate)
-
-	// Register guildCreate as a callback for the guildCreate events.
-	dg.AddHandler(guildCreate)
-
-	// Open the websocket and begin listening.
-	err = dg.Open()
-	if err != nil {
-		fmt.Println("Error opening Discord session: ", err)
-	}
-
-	// Wait here until CTRL-C or other term signal is received.
-	fmt.Println("GoRadBot is now running. Horray! Press CTRL-C to exit.")
-	sc := make(chan os.Signal, 1)
-	signal.Notify(sc, syscall.SIGINT, syscall.SIGTERM, os.Interrupt, os.Kill)
-	<-sc
-
-	// Cleanly close down the Discord session.
-	dg.Close()
-}
